Remove build/root/usr/local/bin/entrypoint (#36)
* Remove build/root/usr/local/bin/entrypoint

In the Prow CI environment, the container's ENTRYPOINT is never
executed (so the Pod is launched with `spec.containers.command` set),
so no reason to keep this `entrypoint` file here.

Instead, set `INSIDE_CI_IMAGE=y` in the Dockerfile and inside `run`
command, print out all the relevant context information:
- git show
- oc version
- $0 $@ of the `run script`

and activate `set -x` to know for sure how `ansible` is launched.

* roles/check_deps/tasks/main.yml: check if oc is there, don't check kubectl if yes

* build/root/usr/local/bin/run: remove while loop

* roles/nv_gpu/tasks/install.yml: give more delay for the NFD CRD to be applied

* roles/nv_gpu/tasks: add verbose_failure.yml to better understand GPU Operator CI failures

* .gitignore: ignore playbooks/gpu_machineset_{definition,aws_template}.json generated files