Prepare artifact directory in toolbox/_common.sh
This patches changes the way CI `run` script and the `toolbox` store
their file artifacts: now, the `toolbox` scripts create a subdirectory
inside `$ARTIFACTS_DIR` (or `/tmp/ci-artifacts_$(date +%Y%m%d)` if
empty).

The name of this `toolbox` script directory is
`<HMS>_<script_path>__<script_name>`.

Ansible logs and facts will be stored in this directory, and
`artifact_extra_logs_dir` points directly to it. This makes it easier
to understand what happened during every step of the deployment and
testing.