gpu_operator_deploy_custom_commit: transform into gpu_operator_bundle_from_commit
This commit rewrites the role `gpu_operator_deploy_custom_commit`
(that was based on the GPU Operator helm charts) and turn it into
`gpu_operator_bundle_from_commit`, which builds the operator (same as
before) from the given commit, but then prepares a bundle image.

I wanted to store this bundle image inside the cluster, so that this
role can run standalone, but it turned out to be impossible at the
moment, because 1/ `operator-sdk` fails to deploy from an ImageStream,
2/ the image must be available locally for the `operator-sdk` process
to access it, but also cluster-wide, so that a `catalog` can be
created to reference it (so a pod-local registry doesn't work).

Because of these constraints that I couldn't all solve, I decided to
rely a quay.io image registry, and hence the script requires quay.io
push credentials to work.

Note that contrary to the original script, this one only builds a
bundle (with the operator image pointing inside the
cluster). `toolbox/gpu-operator/deploy_from_operatorhub.sh` must be
called afterwards, pointing to the quay.io bundle image.