gpu_operator_deploy_from_operatorhub: wait for CRD instead of waiting for CR creating
This commit improves the separation between the initial deployment of
the operator (done by OLM) and the following phase, where the
ClusterPolicy CR is instantiated.

We now explicitly wait for the crd/clusterpolicy.nvidia.com to
appear (OLM deployment),

and then instantiate the ClusterPolicy CR (beginning of the operands
deployment).