CI: GPU Operator: cleanup_cluster between OperatorHub and master branch testing
This is required when testing OperatorHub --> master branch.

Without this cleanup, the entitlement MachineConfig may be undeployed,
but the files still present in the node for several minutes (and gone
when the driver container needs it).

The GPU node scale-down will be required for metrics testing.

I will rename the CI step from 'undeploy_operatorhub' to
'cleanup_cluster' once this is merged.