[deploy-cluster] Add ability to export the metadata.json to a safe place
This commit adds an optional variable flag to export the cluster's
`metadata.json` before the cluster creation.

This is required in the CI to save the file to the shared directory
_before_ creating any resource. So if the cluster creation pod is
interrupted (and the `trap` isn't executed), the `metadata.json` is
still shared with the cluster destruction pod, and the cluster can be
properly destroyed.