cluster_deploy_nginx_server: new role
This role takes a directory in parameter (in addition to the
namespace), which contains the files to server over HTTP.

The role generates a file named `hostname` in its artifact directory,
containing the address at which the HTTP server is listening.

No HTTPS is possible at the moment, this is much more complex to set
up.