rhods_test_jupyterlab: add calls to reset_prometheus_db
Moved this call from _before_ the beginning of the
`rhods_test_jupyterlab` role to right before the creation of the test
job, so that the delay between the reset and the actual test creation
is minimal. In particular, the pre-fetching of the image of the Pod
takes a significant time.