testing: run: download the PR and its comment into the artifacts
The PR json file contains the title, description, commit and base
commit, which are useful information about the version of code being
tested.

The comments of the PR contain, among other things, the `/test`
command that was used to trigger the testing. Soon we'll use this to run
different configurations of the test, eg:
```
/test ods-jh-on-ocp TEST=1 VALUE=2
```