testing: lint and docs improvements
This commit resolves the following lint rules:

- 'yaml'
- 'yaml[truthy]'
- 'risky-file-permissions'
- 'no-changed-when'

It moves 'yaml' to 'yaml[indentation]' and 'yaml[line-length]'
to avoid skipping rules that are needed.

Adds an exclusion folder for the .github folder due to
https://github.com/ansible/ansible-lint/issues/3645

Updates the toolbox repo 'generate_ansible_default_settings'
option to enforce the linting of the parameters and constants.