Include the assets to build a topsail Python package
This commit includes the resource required for
the creation of a Python package called topsail
with the same features and without introducing
any directory structure change change to the current
interfaces calling diretly the run_toolbox.py script.
This commit includes the scripts and jobs so with any
change to the package version there will be a new
released pypi package.

The list of what this commit intends:
- Create the pypi workflow for releasing topsail as a python package in pypi.
- Including the setuptools dependencies to build a python package.
- Organizing the files so the dependencies can be reached from within the
package (running run_toolbox.py <params> and topsail <params> should be equivalent).