Include the assets to build a topsail Python package
This commit includes resource required for
the creation of a Python package called topsail
with the same features and without introducing
any directory structe change change to the current
interfaces calling diretly the run_toolbox.py script.
This commit includes the scripts and jobs so with any
change to the package version there will be a new
released pypi package.